Delphi/all-in-one_client/python/requirements.txt
keyboard==0.13.5
paho-mqtt==1.6.1
Pillow==10.0.1
pip==23.3.1
setuptools==68.2.2
wheel==0.41.2
